Add Movie Service

+++ server/router/index.js
@@ -1,11 +1,12 @@
 const router = require('express').Router();
 
 const { Movie, Category, Comment } = require('../models/');
+const { MovieService, CategoryService, CommentService } = require('../services/');
 const { getGeneratedResponse } = require('../utils/');
 
 router.get('/movies', async (req, res) => {
   try {
-    const movies = await Movie.find();
+    const movies = await MovieService.getAll();
 
     return res.status(200).send(
       getGeneratedResponse(true, movies)
@@ -18,7 +19,7 @@ router.get('/movies', async (req, res) => {
 
 router.post('/movies', async (req, res) => {
   try {
-    const movie = await Movie.create(req.body);
+    const movie = await MovieService.create(req.body);
 
     return res.status(201).send(
       getGeneratedResponse(true, movie)
@@ -32,7 +33,7 @@ router.post('/movies', async (req, res) => {
 router.delete('/movies/:movieId', async (req, res) => {
   try {
     const { movieId } = req.params;
-    const movie = await Movie.findByIdAndDelete(movieId);
+    const movie = await MovieService.delete(movieId);
 
     if (!movie) {
       return res.status(404).send(
@@ -42,10 +43,14 @@ router.delete('/movies/:movieId', async (req, res) => {
       );
     }
 
-    const { deletedCount } = await Comment.deleteMany({ movie: movieId });
+    let deleteCountComments = 0;
+    if (movie.comments.length > 0) {
+      const { deletedCount } = await Comment.deleteMany({ movie: movieId });
+      deleteCountComments = deletedCount;
+    }
 
     return res.status(200).send(
-      getGeneratedResponse(true, movie, { deleteCountComments: deletedCount })
+      getGeneratedResponse(true, movie, { deleteCountComments })
     );
   } catch (err) {
     console.log('Error: ', err.message);
@@ -56,11 +61,7 @@ router.delete('/movies/:movieId', async (req, res) => {
 router.put('/movies/:movieId', async (req, res) => {
   try {
     const { movieId } = req.params;
-
-    const update = req.body;
-    const movie = await Movie.findByIdAndUpdate(movieId, update, {
-      new: true,
-    });
+    const movie = await MovieService.update(movieId, req.body);
 
     if (!movie) {
       return res.status(404).send(
@@ -95,7 +96,7 @@ router.post('/categories', async (req, res) => {
 router.post('/movies/:movieId/comments', async (req, res) => {
   try {
     const { movieId } = req.params;
-    const movie = await Movie.findById(movieId);
+    const movie = await MovieService.getOne(movieId);
 
     if (!movie) {
       return res.status(404).send(
